refactor: Separate classes into own files
