+++ client/index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <title>Battleship</title>
-    <script src="./main.js"></script>
+    <script type="module" src="./main.js"></script>
     <style type="text/css">
         body > * {
             box-sizing: border-box;
